chore(apps/api): update daily price change
