+++ apps/api/src/controllers/flt/index.ts
@@ -1,0 +1,7 @@
+import GetFuseLeveragedTokenBySymbol from "./GetFuseLeveragedTokenBySymbol";
+
+const routers = {
+    GetFuseLeveragedTokenBySymbol,
+};
+
+export default routers;
