+++ apps/api/tests/integrations/GetFuseLeveragedTokenUserPositionById.spec.ts
@@ -1,17 +1,17 @@
 import request from "supertest";
-import server from "../src/server";
-import fltService from "../src/services/flts";
+import server from "../../src/server";
+import fltService from "../../src/services/flts";
 
 afterEach(() => {
     // restore the spy created with spyOn
     jest.restoreAllMocks();
 });
 
-describe("GET /v1/chainId/flts/symbol/backings", () => {
+describe("GET /v1/chainId/positions/positionId", () => {
     describe("given unsupported chainId", () => {
         it("should responds 404 not found", async () => {
             const res = await request(server)
-                .get("/v1/4321/flts/ethrise/backings")
+                .get("/v1/4321/positions/test")
                 .set("Accept", "application/json");
             expect(res.status).toBe(404);
             expect(res.body).toStrictEqual({
@@ -30,53 +30,52 @@ describe("GET /v1/chainId/flts/symbol/backings", () => {
     describe("given Binance Smart Chain", () => {
         describe("given failed request to graphql", () => {
             it("should responds 500 internal server error", async () => {
-                const mock = jest.spyOn(
-                    fltService,
-                    "getFuseLeveragedTokenBackingsBySymbol"
-                );
+                const mock = jest.spyOn(fltService, "getUserPositionById");
                 mock.mockImplementation(() => {
                     throw new Error("some error");
                 });
                 const res = await request(server)
-                    .get("/v1/56/flts/bnbrise/backings")
+                    .get("/v1/56/positions/test")
                     .set("Accept", "application/json");
                 expect(res.status).toBe(500);
             });
         });
 
-        describe("given random symbol", () => {
+        describe("given random positionId", () => {
             it("should responds 404 not found", async () => {
                 const res = await request(server)
-                    .get("/v1/56/flts/hohoho/backings")
+                    .get("/v1/56/positions/test")
                     .set("Accept", "application/json");
                 expect(res.status).toBe(404);
                 expect(res.body).toStrictEqual({
                     errors: [
                         {
                             location: "params",
-                            msg: "symbol not found",
-                            param: "symbol",
-                            value: "hohoho",
+                            msg: "positionId not found",
+                            param: "positionId",
+                            value: "test",
                         },
                     ],
                 });
             });
         });
 
-        describe("given BNBRISE", () => {
+        describe("given correct positionId", () => {
             it("should responds 200 OK", async () => {
                 const res = await request(server)
-                    .get("/v1/56/flts/bnbrise/backings")
+                    .get(
+                        "/v1/56/positions/0x1418be4753a22b69b613fa8b8144d856c023d46b-0x4f7255178b8f15c2cbe92d09b8a77b53ef4ec9ea"
+                    )
                     .set("Accept", "application/json");
                 expect(res.status).toBe(200);
 
-                // Check price
-                const backings = res.body.backings;
-                expect(backings.length).toBeGreaterThan(1);
-                expect(backings[0].timestamp).toBeGreaterThan(10000);
-                expect(typeof backings[0].collateralPerShare).toBe("number");
-                expect(typeof backings[0].debtPerShare).toBe("number");
-                expect(typeof backings[0].valueUSD).toBe("number");
+                // Check user position
+                const position = res.body;
+                expect(position.balance).toBeGreaterThan(5);
+                expect(position.usd).toBeGreaterThan(10);
+                expect(typeof position.pnlUSD).toBe("number");
+                expect(position.pnlPercent).toBeLessThan(100);
+                expect(position.pnlPercent).toBeGreaterThan(-100);
             });
         });
     });
